requirements.txt
fastapi
uvicorn
PyPDF2
openai
faiss-cpu
langchain
pydantic
langchain-community
tiktoken
aiofiles
asyncio
scikit-learn